requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
prefect
streamlit
plotly